requirements.txt
torch
torchvision
transformers
streamlit
pillow
scikit-learn
numpy
matplotlib
tqdm
seaborn
pytorch-grad-cam
huggingface_hub
